Add CI configuration (#17)
* Add CI configuration

This commit adds both a local test and lint configuration for running
tests with tox. A user just needs to install tox locally and run `tox`
and it will run tests and lint locally. A travis configuration is also
added to run tox in CI. The use of tox both locally and in CI helps to
ensure that the same environment (in isolation) is used everywhere.

* Add docs build jobs to tox and travis

This commit adds an initial docs build job to tox and travis. Users can
just run 'tox -edocs' locally and it will build the docs. The travis
job will verify that the docs build without error, but don't host the
output anywhere. The docs build only covers the qiskit_ode.converters
module at this point since the other modules need docstring expansion
before they can be built.