Update pylint version and use black for formatting (#87)
This commit makes a few changes to the code formatting and style rules
used in qiskit-ode. The first thing it does is update the pylint version
to the latest release 2.7.1 (which will match qiskit-terra) which
enables running on python 3.9. The second change is that instead of just
enforcing code style and formatting with pycodestyle and expecting users
to manually adhere to it this uses black instead. Black is an
autoformatter that makes updating the code formatting simple by just
running one command. This takes the guessing out of it and by just
running one fast command you can update all the code formatting to
conform to the rules. Pylint is still run in ci and for the lint tox job
because it does some deeper analysis than just code formatting.

A CONTRIBUTING.md is also added to document all these aspects of working
with the code. It is copy and pasted from the qiskit-experiments repo so
it may not reflect the current state of the repository yet.