Set default branch in reno config
Now that there is stable branch reno is trying to scan the history of
those branches. However, by default reno assumes the base branch is
named "master". On qiskit-dynamics our default branch is named "main" so
reno fails while scanning the history. To fix this issue this commit
sets the default branch to be correctly set to main in the reno config
file. It also bumps the minimum reno version in the requirements-dev.txt
to the 3.4.0 release which was the release which added this
configuration option.