[Stable] Add missing template files to docs configuration (#45) (#47)
* Add missing template files to docs configuration (#45)

* Add missing template files to docs configuration

This commit adds the missing template files to the docs configuration.
This will fix some formatting issues and bring the top navigation bar
inline with the main qiskit documentation.

* Fix docs syntax issues

(cherry picked from commit 6e6f4439c70ae95e678244781a33b98bfd05e3cc)

* Set default branch in reno config

Now that there is stable branch reno is trying to scan the history of
those branches. However, by default reno assumes the base branch is
named "master". On qiskit-dynamics our default branch is named "main" so
reno fails while scanning the history. To fix this issue this commit
sets the default branch to be correctly set to main in the reno config
file. It also bumps the minimum reno version in the requirements-dev.txt
to the 3.4.0 release which was the release which added this
configuration option.