Fix CI by adjusting requirements-dev.txt pinning
Qiskit Terra 0.19.0 released recently which changed the required
matplotlib minimum version and we no longer need to cap < 3.5. To
unblock CI this commit fixes this issue. The other issue with our
dev requirements is we were running black unpinned. Black explicitly
changes the output formatting between releases and if we're relying on
it for CI testing we need to use a fixed version so that CI and local
testing are always using the same version. A recent black release
changed the formatting and broke CI as it expects a different output
formatting than what is in the repo currently. This commit pins the
version to the last working release and we'll just use that version
moving forward until we have a reason to upgrade and update the
formatting in the repo.