Changing `DiscreteSignal.envelope` to return `0.` for times outside sample range (#91)
Originally, DiscreteSignal.envelope would compute the sample to use by
clipping (t - start_time) // dt into [0, len(samples)]. The result was
that for all times before start_time, the envelope would evaluate to 
the first sample, and for all times after the end, it would be the last sample.
This PR changes this so that the envelope alternatively returns 0. outside of 
the sample range.