Added tests + new feature for loading control_channel_map
It is now possible to automatically load a control_channel_map from an existing backend when declaring a DynamicsBackend with the from_backend() method. 

Tests of this features, as well as tests for the previous bug fix have been implemented.

For the new feature, note that manually setting the option control_channel_map of the backend will override the automatic loading of the control_channel_map (when DynamicsBackend is generated with from_backend() method)